requirements.txt
azure-cognitiveservices-speech==1.22.0
chardet==4.0.0
